Added update sidemenu
